+++ shared/nexis/Pages/Dashboard/health_score_tile.h
@@ -35,7 +35,6 @@ private:
 
     HealthScoreCalculator mCalculator;
 
-    QLabel *mLblTitle;
     QLabel *mLblScore;
     QLabel *mLblScoreLabel;